Testing netcdf writing - going to back this out now that ive figured out some stuff with hdf.  Also made shakemap xml grid loader read input data as 32bit float

+++ grid/shake.py
@@ -155,7 +155,7 @@ def readShakeFile(fileobj):
 
     #use the numpy loadtxt function to read in the actual data
     #we're cheating by telling numpy.loadtxt that the last two lines of the XML file are comments
-    data = np.loadtxt(fileobj,comments='<') 
+    data = np.loadtxt(fileobj,comments='<').astype('float32')
     data = data[:,2:] #throw away lat/lon columns
     for i in range(0,len(fields)):
         field = fields[i]
